requirements/base.txt
aioboto3==9.6.0
asgi-correlation-id==3.0.1
asyncpg==0.25.0
cryptography==37.0.4
fastapi==0.78.0
fastapi-cache2[redis]==0.1.8
faust-streaming==0.8.5
httpx==0.23.0
nanoid==2.0.0
prometheus-client==0.14.1
protobuf<=3.20.1
pydantic[email]==1.9.2
pyowm==3.3.0
python-jose==3.3.0
python-json-logger==2.0.4
tortoise-orm==0.19.2
uszipcode==1.0.1
uvicorn[standard]==0.18.2
